exaspoon/exaspoon_next/agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "exaspoon-agents"
version = "0.1.0"
description = "Multi-agent runtime for the EXASPOON finance control surface"
readme = "README.md"
requires-python = ">=3.12.0,<=3.14"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "dateparser>=1.2.0",
    "requests>=2.32.3",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "spoon-ai-sdk",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[tool.uv.sources]
spoon-ai-sdk = { git = "https://github.com/XSpoonAi/spoon-core.git", rev = "main" }
